+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Next release
 
 ### New features
 
+* [#245](https://github.com/statnett/Talk2PowerSystem_PM/issues/245): Allow the chat bot to run with OpenAI LLM not just with Azure
 * [#138](https://github.com/statnett/Talk2PowerSystem_PM/issues/138): The `now` tool now returns the date time in the user's local time zone (using a context var),
 so that time series questions referencing certain point in time are relative to this time (instead of UTC), while the timestamps in Cognite are in UTC.
 Also, modify the agent prompt, so that the agent is aware of that.